requirements.txt
Flask_Login==0.6.3
Werkzeug==3.0.0
Flask==3.0.0
flask_sqlalchemy==3.1.1
SQLAlchemy==2.0.25
python-dotenv~=1.0.1
WTForms~=3.1.2